#46 Fix indexing changelog files from cms-core extension
The indexing script was previously validating only in the main package directory for the presence of an `objects.inv.json` file. Only directories containing this file were scanned and indexed. However, some packages like `cms-core` only have an `objects.inv` file, not `objects.inv.json`.

With this change, the existence of either `objects.inv` or `objects.inv.json` is checked to determine if a directory should be indexed.

Additionally, the list of excluded subfolders has been expanded to include `_panels_static`.