[TASK] Do not exclude Sitemap path from indexing, support x-typo3-indexer meta tag
With this commit the `Sitemap` was removed from the Finder's `noPath` settings.
This will allow files like `XmlSitemap` or files inside `Sitemap` folders to be indexed.

While iterating over the files to be indexed, the indexer will now check if the file
contains an `x-typo3-indexer` meta tag with the value set to `noindex`.
If such a meta tag exists inside the file, the entire file will not be indexed.
This will easily exclude individual files from being indexed.